+++ neo/framework/DeclFX.cpp
@@ -408,7 +408,7 @@ void idDeclFX::ParseSingleFXAction( idLexer &src, idFXSingleAction& FXAction ) {
 idDeclFX::Parse
 ================
 */
-bool idDeclFX::Parse( const char *text, const int textLength, bool allowBinaryVersion ) {
+bool idDeclFX::Parse( const char *text, const size_t textLength, bool allowBinaryVersion /*= false */ ) {
 	idLexer src;
 	idToken token;
