+++ neo/renderer/Model_md5.cpp
@@ -1408,8 +1408,8 @@ void idRenderModelMD5::PurgeModel() {
 idRenderModelMD5::Memory
 ===================
 */
-int	idRenderModelMD5::Memory() const {
-	int total = sizeof( *this );
+size_t	idRenderModelMD5::Memory() const {
+	size_t total = sizeof( *this );
 	total += joints.MemoryUsed() + defaultPose.MemoryUsed() + meshes.MemoryUsed();
 
 	// count up strings
