+++ neo/sys/sys_public.h
@@ -365,7 +365,7 @@ struct sysEvent_t {
 	sysEventType_t	evType;
 	int				evValue;
 	int				evValue2;
-	int				evPtrLength;		// bytes of data pointed to by evPtr, for journaling
+	size_t			evPtrLength;		// bytes of data pointed to by evPtr, for journaling
 	void *			evPtr;				// this must be manually freed if not NULL
 
 	int				inputDevice;
@@ -390,7 +390,7 @@ struct sysMemoryStats_t {
 	int availExtendedVirtual;
 };
 
-typedef unsigned long address_t;
+typedef uintptr_t address_t;
 
 void			Sys_Init();
 void			Sys_Shutdown();
@@ -491,9 +491,9 @@ const char *	Sys_GetCallStackCurAddressStr( int depth );
 void			Sys_ShutdownSymbols();
 
 // DLL loading, the path should be a fully qualified OS path to the DLL file to be loaded
-int				Sys_DLL_Load( const char *dllName );
-void *			Sys_DLL_GetProcAddress( int dllHandle, const char *procName );
-void			Sys_DLL_Unload( int dllHandle );
+ptrdiff_t		Sys_DLL_Load( const char *dllName );
+void *			Sys_DLL_GetProcAddress( ptrdiff_t dllHandle, const char *procName );
+void			Sys_DLL_Unload( ptrdiff_t dllHandle );
 
 // event generation
 void			Sys_GenerateEvents();
