+++ neo/ui/Window.h
@@ -106,7 +106,7 @@ typedef enum {
 
 typedef struct {
 	wexpOpType_t opType;	
-	int	a, b, c, d;
+	ptrdiff_t	a, b, c, d;
 } wexpOp_t;
 
 struct idRegEntry {
@@ -345,9 +345,9 @@ protected:
 	bool RunTimeEvents(int time);
 	void Dump();
 
-	int ExpressionTemporary();
+	ptrdiff_t ExpressionTemporary();
 	wexpOp_t *ExpressionOp();
-	int EmitOp( int a, int b, wexpOpType_t opType, wexpOp_t **opp = NULL );
+	int EmitOp( ptrdiff_t a, ptrdiff_t b, wexpOpType_t opType, wexpOp_t **opp = NULL );
 	int ParseEmitOp( idTokenParser *src, int a, wexpOpType_t opType, int priority, wexpOp_t **opp = NULL );
 	int ParseTerm( idTokenParser *src, idWinVar *var = NULL, int component = 0 );
 	int ParseExpressionPriority( idTokenParser *src, int priority, idWinVar *var = NULL, int component = 0 );
